Make REST Client work whether or not SSL is enabled or not in native
This is done in order to enable the use of Apache HTTP Client in all
cases where the REST Client is used.
This enables us to drop the fallback to URLConnection which is rather limited

Fixes: #9342
