+++ extensions/rest-client/runtime/src/main/java/io/quarkus/restclient/runtime/RestClientRecorder.java
@@ -23,7 +23,6 @@ import io.quarkus.runtime.annotations.Recorder;
 public class RestClientRecorder {
 
     public static ResteasyProviderFactory providerFactory;
-    public static boolean SSL_ENABLED;
 
     public BeanContainerListener hackAroundExtension() {
         return new BeanContainerListener() {
@@ -42,11 +41,6 @@ public class RestClientRecorder {
 
     public void setRestClientBuilderResolver() {
         RestClientBuilderResolver.setInstance(new BuilderResolver());
-    }
-
-    public void setSslEnabled(boolean sslEnabled) {
-        SSL_ENABLED = sslEnabled;
-        RestClientBuilderImpl.setSslEnabled(sslEnabled);
     }
 
     public void initializeResteasyProviderFactory(RuntimeValue<InjectorFactory> ifc, boolean useBuiltIn,
