+++ extensions/rest-client/runtime/src/main/java/io/quarkus/restclient/runtime/graal/ClientBuilderReplacement.java
@@ -3,7 +3,6 @@ package io.quarkus.restclient.runtime.graal;
 import javax.ws.rs.client.ClientBuilder;
 
 import org.jboss.resteasy.client.jaxrs.ResteasyClientBuilder;
-import org.jboss.resteasy.client.jaxrs.engines.URLConnectionClientEngineBuilder;
 import org.jboss.resteasy.client.jaxrs.internal.LocalResteasyProviderFactory;
 import org.jboss.resteasy.client.jaxrs.internal.ResteasyClientBuilderImpl;
 
@@ -19,12 +18,6 @@ final class ClientBuilderReplacement {
     public static ClientBuilder newBuilder() {
         ResteasyClientBuilder client = new ResteasyClientBuilderImpl();
         client.providerFactory(new LocalResteasyProviderFactory(RestClientRecorder.providerFactory));
-        if (!RestClientRecorder.SSL_ENABLED) {
-            client.httpEngine(new URLConnectionClientEngineBuilder().resteasyClientBuilder(client).build());
-            client.sslContext(null);
-            client.trustStore(null);
-            client.keyStore(null, "");
-        }
         return client;
     }
 }
